Sun, Apr 24, 2022, 10:15 AM +00:00

+++ filter-less/helpers.c
@@ -10,9 +10,9 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
         //each column.
         for (int j = 0; j < width; j++)
         {
-            int Red = image[i][j].rgbtRed;
-            int Green = image[i][j].rgbtGreen;
-            int Blue = image[i][j].rgbtBlue;
+            float Red = image[i][j].rgbtRed;
+            float Green = image[i][j].rgbtGreen;
+            float Blue = image[i][j].rgbtBlue;
 
             //Find average.
             int average = round((Red + Green + Blue) / 3);
